+++ packages/protocol/src/protocol/ProtocolEnvironment.ts
@@ -1,6 +1,6 @@
 import { AreProofsEnabled } from "@proto-kit/common";
 
-import { SimpleAsyncStateService } from "../state/StateService";
+import { SimpleAsyncStateService } from "../state/SimpleAsyncStateService";
 import { StateServiceProvider } from "../state/StateServiceProvider";
 
 export interface ProtocolEnvironment {
